feat: display correctly superscript in editorTemplate for enableText elements
